Split declared env executable tests
