+++ docs/COMPLETION_AUDIT.md
@@ -3703,6 +3703,10 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/cli_build/legacy_graph_command_host_effects/file_reads/`,
   preserving deterministic-effect coverage while avoiding one large mixed test
   file.
+- Declared env-read executable graph tests now split normal
+  `zen build build.zen` coverage from direct `zen build.zen` coverage under
+  `tests/integration/cli_build/declared_env_effects/executable/`, preserving
+  the same positive coverage while keeping entrypoint assertions focused.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
